tradecore/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
black==21.5b1
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
more-itertools==8.8.0
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
psycopg2-binary==2.8.3
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5
whitenoise==5.0.1
